refactor(logger): split middleware boundaries
